Refactor CEL variables and add tests
The CloudEvents SDK seems unable to return a single object with both the
standard fields and the extension fields. Until that's possible, the
parsed extensions will be prefixed with `ext.` in the CEL expression.

The parsed JSON data will be prefixed with `data.` in the CEL
expression.

The standard fields are currently un-prefixed, but that will probably
need to change because CEL reserves `type`. Currently the CloudEvents
type is called `typ` as a temporary workaround.

Tests now verify that basic expressions on type, source, parsed
extensions, and parsed data filter correctly.