Prefix CloudEvent context keys
`type` is problematic in CEL because it's a reserved word, and in
general it's probably a bad idea to risk collisions with top-level
reserved words.

This change nests all CE keys under `ce.` to avoid the `type` collision
and all future collisions. A proto is defined for these keys to avoid
issues with dynamic typing and make the definition more shareable.

Eventually the ce and ext prefixes could be combined into a single
prefix for future proofing of extensions and simpler user experience.